chore(suite-native): move pin related components out of connect-device module to be ready for global reuse
