+++ suite-native/device/package.json
@@ -22,6 +22,7 @@
         "@suite-native/atoms": "workspace:*",
         "@suite-native/device-mutex": "workspace:*",
         "@suite-native/feature-flags": "workspace:*",
+        "@suite-native/forms": "workspace:*",
         "@suite-native/helpers": "workspace:*",
         "@suite-native/intl": "workspace:*",
         "@suite-native/link": "workspace:*",
@@ -30,6 +31,7 @@
         "@trezor/connect": "workspace:*",
         "@trezor/device-utils": "workspace:*",
         "@trezor/styles": "workspace:*",
+        "@trezor/validation": "workspace:*",
         "lottie-react-native": "^5.1.5",
         "react": "18.2.0",
         "react-native": "0.71.8",
