+++ suite-native/device/src/constants.ts
