+++ suite-native/module-connect-device/tsconfig.json
@@ -9,12 +9,10 @@
         { "path": "../alerts" },
         { "path": "../atoms" },
         { "path": "../device" },
-        { "path": "../forms" },
         { "path": "../intl" },
         { "path": "../link" },
         { "path": "../navigation" },
         { "path": "../../packages/connect" },
-        { "path": "../../packages/styles" },
-        { "path": "../../packages/validation" }
+        { "path": "../../packages/styles" }
     ]
 }
